Address PR review: fix FKC naming, use to_fasta, remove separators, move toy_gmm, clean tests
- Rename 'Feynman-Kac Control' to 'Feynman-Kac Corrector' everywhere
- Clarify SMC (unbiased denoising) vs FKC (biased by reward) in README
- Remove '(SMC only)' from ess_threshold docs (FKC uses it too)
- Replace manual residue loop with topology.to_fasta() in RMSD CV
- Remove decorative # === separator comment blocks
- Move toy_gmm.py from src/bioemu/ to notebooks/ (not core library)
- Remove trivial/redundant tests, tighten loose threshold (10.0 -> 1.0)
