+++ notebooks/toy_gmm.py
